refactor: separate api routes into modules #14
